+++ packages/babel-jest/src/index.ts
@@ -173,10 +173,11 @@ export const createTransformer: TransformerCreator<
     filename: string,
     transformOptions: JestTransformOptions,
   ): TransformOptions {
-    const {cwd} = transformOptions.config;
-    // `cwd` first to allow incoming options to override it
+    const {cwd, rootDir} = transformOptions.config;
+    // `cwd` and `root` first to allow incoming options to override it
     return {
       cwd,
+      root: rootDir,
       ...options,
       caller: {
         ...options.caller,